refactored pages
