+++ src/app/(state-management)/mobx/post/[id]/page.tsx
@@ -1,4 +1,4 @@
-import SinglePostMobx from '@/components/atomic-design/organisms/SinglePostMobx';
+import SinglePostMobx from 'components/atomic-design/organisms/SinglePostMobx';
 
 export default async function Page({ params }: { params: { id: string } }) {
   const postId = params.id;
