+++ src/app/(state-management)/redux/edit-post/[id]/page.tsx
@@ -1,4 +1,4 @@
-import EditPostFormRedux from '@/components/atomic-design/organisms/EditPostFormRedux';
+import EditPostFormRedux from 'components/atomic-design/organisms/EditPostFormRedux';
 
 export default async function Page({ params }: { params: { id: string } }) {
   const postId = params.id;
