+++ src/app/(state-management)/redux/post/[id]/page.tsx
@@ -1,4 +1,4 @@
-import SinglePostRedux from '@/components/atomic-design/organisms/SinglePostRedux';
+import SinglePostRedux from 'components/atomic-design/organisms/SinglePostRedux';
 
 export default async function Page({ params }: { params: { id: string } }) {
   const postId = params.id;
