+++ src/app/(suspense boundary)/suspense-with-loadingtsx/page.tsx
@@ -2,7 +2,7 @@ async function getData() {
   //node fetch issues if using localhost, src: https://github.com/node-fetch/node-fetch/issues/1624
   let res;
   try {
-    res = await fetch('/api/data', {
+    res = await fetch('http:localhost:3000/api/data', {
       cache: 'no-cache',
     });
   } catch (e) {
