+++ src/store/state-management/mobx/post/transport-layer.ts
@@ -4,10 +4,10 @@ import { transportLayer } from 'store/posts';
 
 export const postTransportLayer: transportLayer = {
   fetchPosts: async () => {
-    return await fetch('http://localhost:3000/api/data/posts');
+    return await fetch('/api/data/posts');
   },
   updateServer: async (title: string, content: string) => {
-    return await fetch('http://localhost:3000/api/data/posts', {
+    return await fetch('/api/data/posts', {
       method: 'POST',
       body: JSON.stringify({
         title: title,
