+++ src/store/state-management/redux/posts/slice.ts
@@ -36,7 +36,7 @@ const initialState: EntityState<Post> & InitialStateType =
 
 // 3b) Create Thunk actions using createAsyncThunk
 export const fetchPosts = createAsyncThunk('posts/fetchPosts', async () => {
-  const response = await fetch('http://localhost:3000/api/data/posts');
+  const response = await fetch('/api/data/posts');
   const allPosts: Post[] = (await response.json()).allPosts;
   return allPosts;
 });
@@ -44,7 +44,7 @@ export const fetchPosts = createAsyncThunk('posts/fetchPosts', async () => {
 export const addNewPost = createAsyncThunk(
   'posts/addNewPost',
   async (initialPost: { title: string; content: string }) => {
-    const res = await fetch('http://localhost:3000/api/data/posts', {
+    const res = await fetch('/api/data/posts', {
       method: 'POST',
       body: JSON.stringify(initialPost),
       headers: {
