+++ tsconfig.json
@@ -20,6 +20,8 @@
       "store/*":["./src/store/*"],
       "data/*":["./src/data/*"],
       "public/*": ["./public/*"],
+      "utils/*":["./src/utils/*"],
+      "components/*":["./src/components/*"],
       "@/*":["*"],
     },// =======================================================================
       //
